adfs

+++ Model/DAO/PromotionDAO.cs
@@ -5,6 +5,7 @@ using Model.ViewModel;
 using System;
 using PagedList;
 
+
 namespace Model.DAO
 {
     public class PromotionDAO
@@ -21,12 +22,12 @@ namespace Model.DAO
             return db.Promotions.Where(x => x.EndTime > DateTime.Now).OrderByDescending(x => x.StartTime).ToList();
         }
 
-        public Promotion getPromotionById(int promotionID)
+        public Promotion getPromotionById(long promotionID)
         {
             return db.Promotions.SingleOrDefault(x => x.ID_Promotion == promotionID);
         }
 
-        public List<PromotionViewModel> getPromotionProduct(int promotionID, int top)
+        public List<PromotionViewModel> getPromotionProduct(long promotionID, int top)
         {
             var model = from prom in db.Promotions
                         join prod in db.Products
@@ -55,7 +56,6 @@ namespace Model.DAO
                 return model.ToList();
             }    
         }
-
         public IEnumerable<Promotion> ListAllPaging(int page = 1, int pageSize = 4)
         {
             return db.Promotions.OrderByDescending(x => x.Name).ToPagedList(page, pageSize);
@@ -74,9 +74,8 @@ namespace Model.DAO
                 return false;
             }
 
-            
-            
+
+
         }
-        
     }
 }
